+++ system/shadowdark.mjs
@@ -1,6 +1,5 @@
 import CompendiumsSD from "./src/documents/CompendiumsSD.mjs";
 import loadTemplates from "./src/templates.mjs";
-// import log from "./src/utils/logging.mjs";
 import Logger from "./src/utils/Logger.mjs";
 import performDataMigration from "./src/migration.mjs";
 import registerHandlebarsHelpers from "./src/handlebars.mjs";
@@ -39,7 +38,6 @@ globalThis.shadowdark = {
 	documents,
 	error: Logger.error,
 	log: Logger.log,
-	logger: Logger,
 	sheets,
 	utils: UtilitySD,
 	warn: Logger.warn,